changed to decimal values for price

+++ bamazon.sql
@@ -8,7 +8,7 @@ CREATE TABLE products (
     item_id INT AUTO_INCREMENT NOT NULL,
     product_name VARCHAR(50) NOT NULL,
     department_name VARCHAR(50) NOT NULL,
-    price INT(10) NOT NULL,
+    price DECIMAL(7,2) NOT NULL,
     stock_quantity INT(10) NOT NULL,
     PRIMARY KEY(item_id)
 );
